Refactor: Add boundary installer download link (#2610)
