--- src/views/product-downloads-view/boundary/components/desktop-client-callout/desktop-client-callout.module.css
@@ -1,34 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-.card {
-	margin-bottom: 32px;
-}
-
-.headingContainer {
-	margin-bottom: 16px;
-}
-
-.downloadContainer {
-	display: grid;
-	gap: 24px;
-	@media (--dev-dot-tablet-up) {
-		grid-template-columns: 1fr 1fr;
-		gap: 28px 24px;
-	}
-}
-
-.downloadCard {
-	display: grid;
-	grid-template-columns: 16px 1fr 24px;
-	align-items: center;
-	column-gap: 8px;
-	&:last-child {
-		margin-bottom: 0;
-	}
-	@media (--dev-dot-tablet-up) {
-		grid-template-columns: 16px 1fr 90px;
-	}
-}
