+++ src/views/product-downloads-view/boundary/components/helpers.ts
