+++ src/views/product-downloads-view/boundary/components/index.tsx
@@ -4,6 +4,5 @@
  */
 
 export * from './icon-linux-color-16'
-export * from './desktop-client-callout'
-export * from './boundary-installer-callout'
+export * from './install-callout'
 export * from './operating-system-icons'
