+++ src/views/product-downloads-view/boundary/download.module.css
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+.lastCard {
+	margin-top: 32px;
+}
